requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
pydantic-settings==2.13.1
sqlalchemy==2.0.49
pandas==3.0.2
openpyxl==3.1.5
python-multipart==0.0.26
httpx==0.28.1
python-dotenv==1.2.2
pytest==9.0.3
deepl>=1.18.0
openai>=1.99.0
